requirements.txt
google-cloud-bigquery
streamlit
db-dtypes
